refactor: 消除 workspace 全部 clippy 警告（121→0）
主要修复类型：
- collapsible_if: 合并嵌套 if 为 let-chain（56 处）
- needless_borrow: 移除不必要的引用（9 处）
- unnecessary_cast: 移除冗余类型转换（12 处）
- too_many_arguments: 添加 #[allow] 属性（6 处）
- let_and_return: 直接返回表达式（3 处）
- 其他：identity_op、useless_conversion、manual_div_ceil、clamp、Range::contains 等

涉及 32 个文件，跨越 backend-wasm、snapshot-format、runtime（含 host_imports 子目录）。

+++ crates/wjsm-runtime/src/host_imports/primitive_core.rs
@@ -603,7 +603,7 @@ pub(crate) fn define_primitive_core(
             } else {
                 value::encode_undefined()
             };
-            write_array_elem(caller, arr_ptr, i as u32, elem);
+            write_array_elem(caller, arr_ptr, i, elem);
         }
         write_array_length(caller, arr_ptr, group_count);
         // .index — 使用 m.start() 保持一致
@@ -659,7 +659,7 @@ pub(crate) fn define_primitive_core(
                     }
                     None => value::encode_undefined(),
                 };
-                write_array_elem(caller, indices_ptr, i as u32, elem);
+                write_array_elem(caller, indices_ptr, i, elem);
             }
             write_array_length(caller, indices_ptr, group_count);
             if !named.is_empty() {
