+++ crates/wjsm-runtime/src/host_imports/string_methods.rs
@@ -53,7 +53,7 @@ pub(crate) fn define_string_methods(
             } else {
                 value::encode_undefined()
             };
-            write_array_elem(caller, arr_ptr, i as u32, elem);
+            write_array_elem(caller, arr_ptr, i, elem);
         }
         write_array_length(caller, arr_ptr, group_count);
         // .index — 使用 m.start() 保持一致
@@ -109,7 +109,7 @@ pub(crate) fn define_string_methods(
                     }
                     None => value::encode_undefined(),
                 };
-                write_array_elem(caller, indices_ptr, i as u32, elem);
+                write_array_elem(caller, indices_ptr, i, elem);
             }
             write_array_length(caller, indices_ptr, group_count);
             if !named.is_empty() {
